feat(server): Start listening to custom events
That way, the application (langgraph) can emit
custom events that the completions endpoint will
receive and process properly.

We are going to use it when the execution is
shortcut because some chit-chatting has been detected
and it's not worth going towards the retrieve and
generate graph, but, instead, just answer.