feat(config): Add support for a new CONTEXTUALISATION_MODEL
If configured, then the query rewrite (context aware, coming
soon) will be executed, providing better contextual questions
and, also, detecting some basic chit-chat cases to avoid
going to the RAG route in a number of situations.

If the variable is not set, than current behaviour remains and
the system will go, always, to the RAG way.