feat(RAG): Introduce context awareness support to the system
This is an opt-in feature (recommended to be enabled) that allows
the RAG system to become context-aware (react differently based
on the history and questions). It provides 2 features:
- Rewrite questions to make them context aware.
- Detect chitchatting questions and avoid routing them via RAG.

Once enabled, the chats with the utility are way more natural
because the searches will be richer (leading to better results)
and, also, the utility will behave better when just chatting
without asking anything.

Note that both features are pretty simple and highly dependent
on the LLM being used (CONTEXTUALISATION_MODEL). Still it's not
able to differentiate Moodle Docs related questions from other
questions, but that's something that we can go progressing with
in the future (analysing search results, reranking, whatever...)