feat(langgraph): Switch to new contexts vs previous configs
This change was introduced by Langgraph 0.6, in preparation
to the incoming (October 2025) release 1.0. Hopefully things
will be more stable after that.

So, with this commit, we are ready for the upcoming version,
at least over the paper. Let's see.

Note that a few pyright ignore comments have been added to a
few lanngchain methods, that's because their typings aren't
ready yet. Again, once version 1.0 is released, we should be
able to remove the exceptions.