fix(langfuse): Ensure that all objects passed are JSON-serializable
Recently pymilvus started to return lists of Hit objects (UserDict),
and those are not JSON-serializable by default. So, or pymilvus
changes them to be serializable or we cast them to normal dicts
that are ok. Here we are using the 2nd approach, that should be
valid for any change in those Hit objects.

Note this was only a problem with Langfuse traces, the system itself
is not affected neither LangSmith traces are.