feat(loader): Modify the dump file format to support metadata
This will be useful to implement the incremental loading, so we
know which was the previous one and can search for changes since
then. Also other meta information about the loaded information can
be added later in the new "meta" property.

Also, no matter it's not implemented yet and looking forward to
add support for multiple mediawiki sites, we are adding all the
site-specific information within a new "sites" list. Obviously, for
now, it will contain only 1 element (1 site).

All the previous pages information now go to a "sites[]/pages" property,
how not. Plug the mediawiki url and the number of pages.

Finally, note that the indexer, when loading the dump files, will
be able to manage both the old and the new formats in a transparent
way, at least for some good time, allowing all the cases to be
updated without trouble.